Fix matching bug
Bug was to use drain iterator on set of to_inspect elements.
It cleared the whole map.

Fixed by manually picking an element and removing it before handling if
successful.

Added a lot of debug code to visualize matching.

Fixed a crash in output_dot_filtered due to coloring.
Code expected some map values to be set, which were not due to
filtering.
Fix is to compute coloring for all elements.
Ultimately is would be better to define mapping of values to graph
objects, or graph partial views.