Add nav bar and static file handling
Add a nav bar for the whole wiki.
Add a small css file for styling the nav bar (nav box on the right).

Add system to try loading an asset file if request not caught by any dynamic API.
Assets will be stored in the binary in release mode.
Assets are returned with a cache indication, as they are static.