Parsing Database from file
Untested, todo restore io tests.

Heavy use of match to do parsing control flow in big steps at once.
This is nicer than expected, the parsing code is very small.

As with all versions, the code for validating the database post-parsing
and restoring the cross-references is the biggest one.
This is in part due to the borrow checker, which does not see that we
manipulate two 'independent' spaces: element values, and element
cross-references.