Pass Request by reference
This allows Page types to be reference wrappers, because a lifetime is specified.
This is used for StaticAsset { name: &str }.
This reduces code in matching, and simplifies the routing function.

However, lifetimes are annoying on functions due to HTRB.
For now this works with a local array of handlers but not a const one.
I will try to improve this.