Remove PathFormat system
This system is nice-looking but suffers from problems:

Some type annotations must be used for parsing result.
These annotations negate the compactness advantage.

It cannot be generic for all number of elements.
An impl for each tuple arity must be provided.

Thus it has been removed in favor of the simple approach.