Restore wiki_page, rename concepts
Renamed Page to EndPoint to better reflect the concept.
StaticAsset is not really a page anyway.

Restored wiki_page function which performed final assembly.
Make use of EndPoint::url() inside it for StaticAsset's.

Unfortunately it is not possible to auto impl horrorshow::Render traits for EndPoint with url().
This would collide with a blanket impl in horrorshow.
Trying to remove the need for ".url()" would cost a lot more than 4 chars...