Improve relations construction
Finally found proper ownership scheme to simplify deserialize, and extract the "registering" step in function.
register_* functions take the element by value, which is ok because we need it by value internally.

Modified database to have a map by type of atom (onyl 1 for now).
Expose function to match directly on this map without copying by converting to Atom.